backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
aiofiles
pytest
python-multipart